did good

+++ libft/libft.h
@@ -35,7 +35,7 @@ char    *ft_strrchr(const char *s, int c);
 void    ft_bzero(void *s, unsigned int n);
 void    *ft_memchr(void *s, int c, unsigned int n);
 void    *ft_memmove(void *dest, void *src, unsigned int n);
-void    *ft_calloc(unsigned int nmemb, unsigned int size);
+void    *ft_calloc(size_t nmemb, size_t size);
 char *ft_substr(char const *s, unsigned int start, size_t len);
 char *ft_strjoin(char const *s1, char const *s2);
 char *ft_strtrim(char const *s1, char const *set);
@@ -56,9 +56,6 @@ void ft_lstdelone(t_list *lst, void (*del)(void*));
 void ft_lstclear(t_list **lst, void (*del)(void*));
 void ft_lstiter(t_list *lst, void (*f)(void *));
 t_list *ft_lstmap(t_list *lst, void *(*f)(void *), void (*del)(void *));
-
-
-
 
 
 #endif
